fix: 변수 컨벤션 및 중요도 낮은 클래스는 inner class로 넣음
